+++ Changelog.md
@@ -7,6 +7,16 @@ This project does _not_ adhere to
 [Semantic Versioning](https://semver.org/spec/v2.0.0.html) but contrary to it
 every new version is a new major version.
 
+## 170.0.0 - 2024-03-25
+
+### Fixed
+
+-   `Feedback` gets reset after closing.
+
+### Removed
+
+-   `Give more feedback` button in `Feedback` dialog.
+
 ## 169.0.0 - 2024-03-25
 
 ### Added